Add -dry_run flag (#275)
Fix #274

+++ shallow_backup/reinstall.py
@@ -11,7 +11,7 @@ from shutil import copytree, copyfile, copy
 #       conflict with the function names.
 
 
-def reinstall_dots_sb(dots_path: str, home_path: str = os.path.expanduser("~")):
+def reinstall_dots_sb(dots_path: str, home_path: str = os.path.expanduser("~"), dry_run: bool = False):
 	"""Reinstall all dotfiles and folders by copying them from dots_path
 	to a path relative to home_path, or to an absolute path."""
 	exit_if_dir_is_empty(dots_path, 'dotfile')
@@ -42,19 +42,21 @@ def reinstall_dots_sb(dots_path: str, home_path: str = os.path.expanduser("~")):
 	# The absolute file paths prepended with ':' are converted back to valid paths
 	# Format: [(source, dest), ... ]
 	full_path_dotfiles_to_reinstall = []
-	for dot in dotfiles_to_reinstall:
-		source = os.path.join(dots_path, dot)
+	for source in dotfiles_to_reinstall:
 		# If it's an absolute path, dest is the corrected path
-		if dot.startswith(":"):
-			dest = "/" + dot[1:]
+		if source.startswith(":"):
+			dest = "/" + source[1:]
 		else:
 			# Otherwise, it should go in a path relative to the home path
-			relative_dotfile_path = dot.replace(dots_path, "")
-			dest = os.path.join(home_path, relative_dotfile_path)
+			dest = source.replace(dots_path, home_path + "/")
 		full_path_dotfiles_to_reinstall.append((Path(source), Path(dest)))
 
 	# Copy files from backup to system
 	for dot_source, dot_dest in full_path_dotfiles_to_reinstall:
+		if dry_run:
+			print_dry_run_copy_info(dot_source, dot_dest)
+			continue
+
 		# Create dest parent dir if it doesn't exist
 		safe_mkdir(dot_dest.parent)
 		try:
@@ -67,20 +69,23 @@ def reinstall_dots_sb(dots_path: str, home_path: str = os.path.expanduser("~")):
 	print_section_header("DOTFILE REINSTALLATION COMPLETED", Fore.BLUE)
 
 
-def reinstall_fonts_sb(fonts_path: str):
+def reinstall_fonts_sb(fonts_path: str, dry_run: bool = False):
 	"""Reinstall all fonts."""
 	exit_if_dir_is_empty(fonts_path, 'font')
 	print_section_header("REINSTALLING FONTS", Fore.BLUE)
 
 	# Copy every file in fonts_path to ~/Library/Fonts
 	for font in get_abs_path_subfiles(fonts_path):
-		font_lib_path = get_fonts_dir()
-		dest_path = os.path.join(font_lib_path, font.split("/")[-1])
+		fonts_dir = get_fonts_dir()
+		dest_path = quote(os.path.join(fonts_dir, font.split("/")[-1]))
+		if dry_run:
+			print_dry_run_copy_info(font, dest_path)
+			continue
 		copyfile(quote(font), quote(dest_path))
 	print_section_header("FONT REINSTALLATION COMPLETED", Fore.BLUE)
 
 
-def reinstall_configs_sb(configs_path: str):
+def reinstall_configs_sb(configs_path: str, dry_run: bool = False):
 	"""Reinstall all configs from the backup."""
 	exit_if_dir_is_empty(configs_path, 'config')
 	print_section_header("REINSTALLING CONFIG FILES", Fore.BLUE)
@@ -88,18 +93,30 @@ def reinstall_configs_sb(configs_path: str):
 	config = get_config()
 	for dest_path, backup_loc in config["config_mapping"].items():
 		dest_path = quote(dest_path)
-		path_to_backup = quote(os.path.join(configs_path, backup_loc))
-		# TODO: REFACTOR WITH GENERIC COPY FUNCTION.
-		if os.path.isdir(path_to_backup):
-			copytree(path_to_backup, dest_path)
-		elif os.path.isfile(path_to_backup):
-			copyfile(path_to_backup, dest_path)
+		source_path = quote(os.path.join(configs_path, backup_loc))
+
+		if dry_run:
+			print_dry_run_copy_info(source_path, dest_path)
+			continue
+
+		if os.path.isdir(source_path):
+			copytree(source_path, dest_path)
+		elif os.path.isfile(source_path):
+			copyfile(source_path, dest_path)
 
 	print_section_header("CONFIG REINSTALLATION COMPLETED", Fore.BLUE)
 
 
-def reinstall_packages_sb(packages_path: str):
+def reinstall_packages_sb(packages_path: str, dry_run: bool = False):
 	"""Reinstall all packages from the files in backup/installs."""
+	def run_cmd_if_no_dry_run(command, dry_run) -> int:
+		if dry_run:
+			print_yellow_bold(f"$ {command}")
+			# Return 0 for any processes depending on chained successful commands
+			return 0
+		else:
+			return run_cmd(command)
+
 	exit_if_dir_is_empty(packages_path, 'package')
 	print_section_header("REINSTALLING PACKAGES", Fore.BLUE)
 
@@ -121,30 +138,30 @@ def reinstall_packages_sb(packages_path: str):
 		if pm in ["brew", "brew-cask"]:
 			pm_formatted = pm.replace("-", " ")
 			print_pkg_mgr_reinstall(pm_formatted)
-			cmd = "xargs {0} install < {1}/{2}_list.txt".format(pm.replace("-", " "), packages_path, pm_formatted)
-			run_cmd(cmd)
+			cmd = f"xargs {pm.replace('-', ' ')} install < {packages_path}/{pm_formatted}_list.txt"
+			run_cmd_if_no_dry_run(cmd, dry_run)
 		elif pm == "npm":
 			print_pkg_mgr_reinstall(pm)
-			cmd = "cat {0}/npm_list.txt | xargs npm install -g".format(packages_path)
-			run_cmd(cmd)
+			cmd = f"cat {packages_path}/npm_list.txt | xargs npm install -g"
+			run_cmd_if_no_dry_run(cmd, dry_run)
 		elif pm == "pip":
 			print_pkg_mgr_reinstall(pm)
-			cmd = "pip install -r {0}/pip_list.txt".format(packages_path)
-			run_cmd(cmd)
+			cmd = f"pip install -r {packages_path}/pip_list.txt"
+			run_cmd_if_no_dry_run(cmd, dry_run)
 		elif pm == "pip3":
 			print_pkg_mgr_reinstall(pm)
-			cmd = "pip3 install -r {0}/pip3_list.txt".format(packages_path)
-			run_cmd(cmd)
+			cmd = f"pip3 install -r {packages_path}/pip3_list.txt"
+			run_cmd_if_no_dry_run(cmd, dry_run)
 		elif pm == "vscode":
 			print_pkg_mgr_reinstall(pm)
-			with open("{0}/vscode_list.txt".format(packages_path), "r") as f:
-				for x in f:
-					cmd = "code --install-extension {0}".format(x)
-					run_cmd(cmd)
+			with open(f"{packages_path}/vscode_list.txt", "r") as file:
+				for package in file:
+					cmd = f"code --install-extension {package}"
+					run_cmd_if_no_dry_run(cmd, dry_run)
 		elif pm == "apm":
 			print_pkg_mgr_reinstall(pm)
-			cmd = "apm install --packages-file {0}/apm_list.txt".format(packages_path)
-			run_cmd(cmd)
+			cmd = f"apm install --packages-file {packages_path}/apm_list.txt"
+			run_cmd_if_no_dry_run(cmd, dry_run)
 		elif pm == "macports":
 			print_red_bold("WARNING: Macports reinstallation is not supported.")
 		elif pm == "gem":
@@ -156,9 +173,9 @@ def reinstall_packages_sb(packages_path: str):
 	print_section_header("PACKAGE REINSTALLATION COMPLETED", Fore.BLUE)
 
 
-def reinstall_all_sb(dotfiles_path: str, packages_path: str, fonts_path: str, configs_path: str):
+def reinstall_all_sb(dotfiles_path: str, packages_path: str, fonts_path: str, configs_path: str, dry_run: bool = False):
 	"""Call all reinstallation methods."""
-	reinstall_dots_sb(dotfiles_path)
-	reinstall_packages_sb(packages_path)
-	reinstall_fonts_sb(fonts_path)
-	reinstall_configs_sb(configs_path)
+	reinstall_dots_sb(dotfiles_path, dry_run=dry_run)
+	reinstall_packages_sb(packages_path, dry_run=dry_run)
+	reinstall_fonts_sb(fonts_path, dry_run=dry_run)
+	reinstall_configs_sb(configs_path, dry_run=dry_run)
